Fixing bunch of issues:
=> Apollo codegen, apollo run time and all plugins updated to 0.5.0
=> all issues related to upgrade is solved
=> Fixed the deprecated authentication call from yelp, now we don't need to call authorization end point for queries.
=> Updated the schema to latest
=> Fixed all the compile issues and upgraded all dependencies along with java 8 support.

Next update we will make most of the things to kotlin.

+++ app/src/main/java/binarybricks/com/yelpql/network/model/Business.java
@@ -15,7 +15,7 @@ public class Business {
     private String url;
     private String phone;
     private String displayPhone;
-    private Integer reviewCount;
+    private Long reviewCount;
     private String rating;
     private String price;
     private List<String> categories;
@@ -68,11 +68,11 @@ public class Business {
         this.displayPhone = displayPhone;
     }
 
-    public Integer getReviewCount() {
+    public Long getReviewCount() {
         return reviewCount;
     }
 
-    public void setReviewCount(Integer reviewCount) {
+    public void setReviewCount(Long reviewCount) {
         this.reviewCount = reviewCount;
     }
